Added the contract entity and aplied it to all instances which implement it.
